Fix warnings in cpp-tests.

+++ cocos/scripting/lua-bindings/manual/lua_cocos2dx_deprecated.cpp
@@ -2092,7 +2092,7 @@ static int tolua_bnd_cast_deprecated00(lua_State* tolua_S)
     void* v = nullptr;
     std::string strValue = "";
     strValue = tolua_tostring(tolua_S,2,NULL);
-    int pos = strValue.find("CC");
+    size_t pos = strValue.find("CC");
     if (pos == 0 &&
         std::string::npos == strValue.find("CCBAnimationManager") &&
         std::string::npos == strValue.find("CCString") &&
